requirements.txt
asgiref==3.8.1
aspectlib==2.0.0
cffi==1.17.0
cryptography==43.0.0
Django==5.0.7
django-environ==0.11.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
environ==1.0
fields==5.0.0
psycopg==3.2.1
pycparser==2.22
PyJWT==2.9.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
